init_players.py refactored
